Criou as verificações das constantes do banco 06/06/23

+++ backend/src/app/useCases/categories/deleteCategory.ts
@@ -6,6 +6,12 @@ export async function deleteCategory(req: Request, res: Response) {
   try {
     const { categoryId } = req.params;
 
+    if (!categoryId) {
+      return res.status(400).json({
+        error: 'categoryId is required',
+      });
+    }
+
     await Category.findByIdAndDelete(categoryId);
 
     res.sendStatus(204);
